Asset-Assessment-Scanner-V1/requirements.txt
defusedxml==0.7.1
Faker==37.1.0
fonttools==4.57.0
fpdf2==2.8.3
lxml==5.4.0
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
